Split prepare runner responsibilities
